fix(deno): rework imports and change stuff around to get Deno working
