+++ src/methods/skyblock/auction.ts
@@ -1,5 +1,5 @@
 // @deno-types="../../types/api.ts"
-import type { Paths } from "../../types/api";
+import { Paths } from "../../types/api";
 // @deno-types="../../util/Method.ts"
 import { Method } from "../../util/Method";
 // @deno-types="../../util/ResultArray.ts"
