+++ src/util/BaseClient.ts
@@ -1,0 +1,9 @@
+// @deno-types="../types/api.ts"
+import { Components } from "../types/api";
+
+export abstract class BaseClient {
+  public abstract call<T extends Components.Schemas.ApiSuccess>(
+    path: string,
+    parameters?: Record<string, string>
+  ): Promise<T & { cached?: boolean }>;
+}
