+++ src/util/Method.ts
@@ -1,13 +1,13 @@
-// @deno-types="../Client.ts"
-import type { Client } from "../Client";
+// @deno-types="./BaseClient.ts"
+import { BaseClient } from "./BaseClient";
 
 /** @internal */
 export abstract class Method {
   /** @internal */
-  protected client: Client;
+  protected client: BaseClient;
 
   /** @internal */
-  constructor(client: Client) {
+  constructor(client: BaseClient) {
     this.client = client;
   }
 }
