+++ src/util/ResultArray.ts
@@ -1,7 +1,7 @@
-// @deno-types="../Client.ts"
-import type { DefaultMeta } from "../Client";
+// @deno-types="../types/DefaultMeta.ts"
+import { DefaultMeta } from "../types/DefaultMeta";
 // @deno-types="../types/api.ts"
-import type { Components } from "../types/api";
+import { Components } from "../types/api";
 
 /**
  * Generic intersection type for result arrays to include metadata as a non-enumerable property.
